In cells.m, added methods rm_duplicates and patch_empty.
In orbit.m:
- updated the object class to handle sp3 files better;
- added some common parameters to the object parameter list;
- added list of sp3-specific parameters;
- made descriptor-type object parameters to be public;
- added wrapper methods to the data types;
- added translatesatname mehtod, inverse of existing translatesat method;
- updated the load_ascii method to:
  - use the new file methods;
  - handle the sp3k format;
  - be more concise when building the simpletimeseries objects;
- revised most commonly-used methods, update the code and cleaned up some bugs;
- implemented time system wrapers (to the correspondent in simpletimeseries);
- implemented the get method, to be retrieve data types;
- implemented the isempty method;
- implemented the sp3 exporting method;
- revised the import routines to consistently return metadata that is compatible with the sp3 format.

In simpletimeseries.m. fixed minor bug int he set.tsys method.

In str.m, added the trunc method, to truncated strings and handle default values for empty strings.

In time.m, added the method fod (fraction of day).

In varargs.m, corrected the help string of the wrap method.

Added the version_patching directory, that contains routines that are available in other matlab versions but not in the current one.